Changed from "cql" to "cass"
This is going to look massive, but it’s not. It’s just a bunch of
renaming. The only somewhat “meaty” part is in cassandra.cpp. I added
an abstraction to map between internal and external types. This is much
cleaner (except for the casting ugliness in that file).  I went with
“cass” because I couldn’t come up with anything better. I’m open to
suggestions if someone has a better idea. “cassandra” is just too long
IMO. I also don’t think it’s necessary to prefix all the internal stuff
with “cass” (including filenames). It get’s tedious typing that over
and over, with directories and namespaces we don’t
have to.