Token aware refactor and fixes
I broke some stuff in TokenMap when making changes for code review.

Moved the strategy class/options processing out of schema and into
the replication strategy classes themselves.

Remove RTTI for comparing replication strategies. I'm not a big fan.

Renamed "ReplicationPlacementStrategy" to "ReplicationStrategy" because
that's the commonly used name.

Renamed "CassTypeParser" to "TypeParser" because I try to reserve
"Cass*" for stuff that's in the external API.