Added functions with explicit string lengths
Languages that are wrapping the driver might not have null-terminated
strings. This forces them to copying to an extra buffer just to add a
terminator.

Functions with string lengths use the "_n" suffix.
`cass_session_execute()` and `cass_session_prepare()` have been converted
to this new pattern. This is a breaking change, but it makes the API much
more consistent. Also, `cass_string_init2()` has been renamed to
`cass_string_init_n()` for consistency.