test: CPP-150 Latency Aware Policy integration test
This test will ensure latency is incurred on all hosts connected in a
session. A new feature in CCM bridge is being used to send signals
(SIGSTOP/SIGCONT) on POSIX platform or pausing/resuming all threads in a
process on Windows. Since this is not deterministic an acceptable
tolerence is applied for latencies that are less than expected.

This commit also enhances host logging of latency and allow for Boost
location to be specified on the command line through BOOST_ROOT_DIR.

PR: https://github.com/datastax/cpp-driver/pull/153
Reviewed-By: Michael Penick <michael.penick@datastax.com>