CPP-466 - Shuffle replicas when using token aware routing
This enables randomly shuffling replicas when using token aware
routing as opposed to using round-robin with a random start index. This
improves replica distribution for a given partition key during node
failures.

This is feature is enabled by default as it more closely matches the
previous functionality. It can be disabled to improve server-side
caching behavior or for debugging/testing.