CPP-478 - Rename SpeculativeExecution to RequestExecution (#37)
"SpeculativeExecution" implies that it is a special type of execution,
when really it's just one attempt at executing a request.
SpeculativeExecutionPolicy and related classes govern the actions
of "speculative execution", which is simply the act of speculatively
executing a request when an execution is already in progress on a
different node.