Moved to type safe callbacks
Use a type safe callback class instead of using a `void*` data member
and a function pointer. It's way more flexible, testable, and prevents
casting errors when dealing with `void*`.

The `Callback` class is a really limited version of something like
`std::function` without heap allocation. Presently, it only allows for a
single argument, but that's all we seem to need for now.

Also,
* Removed `AsyncQueue` which was no longer used.
* Fixed memory issues in the connection pool unit tests and
  `ConnectionPool`.