CPP-667 - Fix timer handle leaks (#174)
`Timer::stop()` now means "I'm done with this timer stop and clean it
up". The timer can still be restarted using `start()`, but it's
slightly more expensive because it needs to recreate a new timer handle
instead of reusing the old handle.

There were many cases where timer handles were being left open because
`Timer::stop()` was called, but not `Timer::close_handle()`. In some
cases the timer was also closed on the wrong thread.  There are just not
any cases currently that benefit from the handles being left open after
the timer is stopped explicitly using `stop()` so its implementation is
is now that of `close_handle()`.