CPP-687 - Fix recovery of remote DC host (#427)
The DC-aware load balancing policy depends on an internal host map to determine host distance. The problem was the host distance was being checked to determine whether a node should be added or marked as up.  This fix always adds and marks hosts up even if they're ignored. The distance is now only used to determine if a host should have a connection pool.