Fixing local_dc_ assignment and addressing review comments
The local DC is now only being applied to the DCAware policy if the
current `local_dc_` is not set. New unit tests were added to ensure that
the local DC is not overridden when set in the default or execution
profiles and also set when the local DC is not initially set.