Allow remote hosts to come back up
When a host comes back up after being down, a request processor
(unlike a control connection) only notifies the load balancing
policy if the host is not ignored.

Sadly some policies ignore remote hosts that are down.

In combination, this means that remote hosts can never be used by
a request (processor) after having once gone down.

This fix resolves this, by always notifying a policy of all
host-up events, regardless of the distance of that host.