Reach full compatibility with OpenSSL 1.1
Despite what the documentation[1] says the code is not fully compatible
with OpenSSL 1.1 because it relies on compatibility wrappers behind
OPENSSL_API_COMPAT.

The OpenSSL documentation on library initialization[2] explains that
none of the startup fucntions are necessary starting with 1.1. The
changelog for "changes between 1.0.2h and 1.1.0 [25 Aug 2016]"[3]
mentions several removed shutdown functions. The removal of
`CRYPTO_set_locking_callback` is mentioned in a GitHub comment[4].

[1] https://docs.datastax.com/en/developer/cpp-driver/2.15/topics/building/
[2] https://wiki.openssl.org/index.php/Library_Initialization
[3] https://www.openssl.org/news/changelog.html
[4] https://github.com/openssl/openssl/issues/1260#issuecomment-228846202