Remove unreachable code (#535)
* Remove unreachable code

This was noticed by MSVC in a debug build which produced the following
warning:

    warning C4702: unreachable code

* Update slot outside of the loop in MPMCQueue

As suggested by @absurdfarce this change keeps finding a slot in the
loop but moves the use of the slot past the loop. That way there is a
return after the loop which should make clear to any code analysis that
this function will always return something.